Move files and write comments.
