+++ src/Backups/BackupCoordinationReplicatedTables.cpp
@@ -1,10 +1,7 @@
-#include <Backups/BackupCoordinationHelpers.h>
+#include <Backups/BackupCoordinationReplicatedTables.h>
 #include <Storages/MergeTree/MergeTreePartInfo.h>
 #include <Storages/MergeTree/ReplicatedMergeTreeMutationEntry.h>
 #include <Common/Exception.h>
-#include <Common/escapeForFileName.h>
-#include <IO/ReadHelpers.h>
-#include <base/chrono_io.h>
 #include <boost/range/adaptor/map.hpp>
 
 
@@ -14,7 +11,6 @@ namespace DB
 namespace ErrorCodes
 {
     extern const int CANNOT_BACKUP_TABLE;
-    extern const int FAILED_TO_SYNC_BACKUP_OR_RESTORE;
     extern const int LOGICAL_ERROR;
 }
 
@@ -30,7 +26,7 @@ namespace
 using MutationInfo = IBackupCoordination::MutationInfo;
 
 
-class BackupCoordinationReplicatedPartsAndMutations::CoveredPartsFinder
+class BackupCoordinationReplicatedTables::CoveredPartsFinder
 {
 public:
     explicit CoveredPartsFinder(const String & table_name_for_logs_) : table_name_for_logs(table_name_for_logs_) {}
@@ -150,10 +146,10 @@ private:
 };
 
 
-BackupCoordinationReplicatedPartsAndMutations::BackupCoordinationReplicatedPartsAndMutations() = default;
-BackupCoordinationReplicatedPartsAndMutations::~BackupCoordinationReplicatedPartsAndMutations() = default;
+BackupCoordinationReplicatedTables::BackupCoordinationReplicatedTables() = default;
+BackupCoordinationReplicatedTables::~BackupCoordinationReplicatedTables() = default;
 
-void BackupCoordinationReplicatedPartsAndMutations::addPartNames(
+void BackupCoordinationReplicatedTables::addPartNames(
     const String & table_shared_id,
     const String & table_name_for_logs,
     const String & replica_name,
@@ -204,7 +200,7 @@ void BackupCoordinationReplicatedPartsAndMutations::addPartNames(
     }
 }
 
-Strings BackupCoordinationReplicatedPartsAndMutations::getPartNames(const String & table_shared_id, const String & replica_name) const
+Strings BackupCoordinationReplicatedTables::getPartNames(const String & table_shared_id, const String & replica_name) const
 {
     prepare();
     
@@ -220,7 +216,7 @@ Strings BackupCoordinationReplicatedPartsAndMutations::getPartNames(const String
     return it2->second;
 }
 
-void BackupCoordinationReplicatedPartsAndMutations::addMutations(
+void BackupCoordinationReplicatedTables::addMutations(
     const String & table_shared_id,
     const String & table_name_for_logs,
     const String & replica_name,
@@ -239,7 +235,7 @@ void BackupCoordinationReplicatedPartsAndMutations::addMutations(
 }
 
 std::vector<MutationInfo>
-BackupCoordinationReplicatedPartsAndMutations::getMutations(const String & table_shared_id, const String & replica_name) const
+BackupCoordinationReplicatedTables::getMutations(const String & table_shared_id, const String & replica_name) const
 {
     prepare();
     
@@ -257,7 +253,24 @@ BackupCoordinationReplicatedPartsAndMutations::getMutations(const String & table
     return res;
 }
 
-void BackupCoordinationReplicatedPartsAndMutations::prepare() const
+void BackupCoordinationReplicatedTables::addDataPath(const String & table_shared_id, const String & data_path)
+{
+    auto & table_info = table_infos[table_shared_id];
+    table_info.data_paths.emplace(data_path);
+}
+
+Strings BackupCoordinationReplicatedTables::getDataPaths(const String & table_shared_id) const
+{
+    auto it = table_infos.find(table_shared_id);
+    if (it == table_infos.end())
+        return {};
+    
+    const auto & table_info = it->second;
+    return Strings{table_info.data_paths.begin(), table_info.data_paths.end()};
+}
+
+
+void BackupCoordinationReplicatedTables::prepare() const
 {
     if (prepared)
         return;
@@ -317,182 +330,6 @@ void BackupCoordinationReplicatedPartsAndMutations::prepare() const
     }
 
     prepared = true;
-}
-
-
-/// Helper designed to be used in an implementation of the IBackupCoordination interface in the part related to replicated access storages.
-BackupCoordinationReplicatedAccess::BackupCoordinationReplicatedAccess() = default;
-BackupCoordinationReplicatedAccess::~BackupCoordinationReplicatedAccess() = default;
-
-void BackupCoordinationReplicatedAccess::addFilePath(const String & access_zk_path, AccessEntityType access_entity_type, const String & host_id, const String & file_path)
-{
-    auto & ref = file_paths_by_zk_path[std::make_pair(access_zk_path, access_entity_type)];
-    ref.file_paths.emplace(file_path);
-
-    /// std::max() because the calculation must give the same result being repeated on a different replica.
-    ref.host_to_store_access = std::max(ref.host_to_store_access, host_id);
-}
-
-Strings BackupCoordinationReplicatedAccess::getFilePaths(const String & access_zk_path, AccessEntityType access_entity_type, const String & host_id) const
-{
-    auto it = file_paths_by_zk_path.find(std::make_pair(access_zk_path, access_entity_type));
-    if (it == file_paths_by_zk_path.end())
-        return {};
-
-    auto & file_paths = it->second;
-    if (file_paths.host_to_store_access != host_id)
-        return {};
-
-    Strings res{file_paths.file_paths.begin(), file_paths.file_paths.end()};
-    return res;
-}
-
-
-/// Helps to wait until all hosts come to a specified stage.
-BackupCoordinationStatusSync::BackupCoordinationStatusSync(const String & zookeeper_path_, zkutil::GetZooKeeper get_zookeeper_, Poco::Logger * log_)
-    : zookeeper_path(zookeeper_path_)
-    , get_zookeeper(get_zookeeper_)
-    , log(log_)
-{
-    createRootNodes();
-}
-
-void BackupCoordinationStatusSync::createRootNodes()
-{
-    auto zookeeper = get_zookeeper();
-    zookeeper->createAncestors(zookeeper_path);
-    zookeeper->createIfNotExists(zookeeper_path, "");
-}
-
-void BackupCoordinationStatusSync::set(const String & current_host, const String & new_status, const String & message)
-{
-    setImpl(current_host, new_status, message, {}, {});
-}
-
-Strings BackupCoordinationStatusSync::setAndWait(const String & current_host, const String & new_status, const String & message, const Strings & all_hosts)
-{
-    return setImpl(current_host, new_status, message, all_hosts, {});
-}
-
-Strings BackupCoordinationStatusSync::setAndWaitFor(const String & current_host, const String & new_status, const String & message, const Strings & all_hosts, UInt64 timeout_ms)
-{
-    return setImpl(current_host, new_status, message, all_hosts, timeout_ms);
-}
-
-Strings BackupCoordinationStatusSync::setImpl(const String & current_host, const String & new_status, const String & message, const Strings & all_hosts, const std::optional<UInt64> & timeout_ms)
-{
-    /// Put new status to ZooKeeper.
-    auto zookeeper = get_zookeeper();
-    zookeeper->createIfNotExists(zookeeper_path + "/" + current_host + "|" + new_status, message);
-
-    if (all_hosts.empty() || (new_status == kErrorStatus))
-        return {};
-
-    if ((all_hosts.size() == 1) && (all_hosts.front() == current_host))
-        return {message};
-
-    /// Wait for other hosts.
-
-    Strings ready_hosts_results;
-    ready_hosts_results.resize(all_hosts.size());
-
-    std::map<String, std::vector<size_t> /* index in `ready_hosts_results` */> unready_hosts;
-    for (size_t i = 0; i != all_hosts.size(); ++i)
-        unready_hosts[all_hosts[i]].push_back(i);
-
-    std::optional<String> host_with_error;
-    std::optional<String> error_message;
-
-    /// Process ZooKeeper's nodes and set `all_hosts_ready` or `unready_host` or `error_message`.
-    auto process_zk_nodes = [&](const Strings & zk_nodes)
-    {
-        for (const String & zk_node : zk_nodes)
-        {
-            if (zk_node.starts_with("remove_watch-"))
-                continue;
-
-            size_t separator_pos = zk_node.find('|');
-            if (separator_pos == String::npos)
-                throw Exception(ErrorCodes::FAILED_TO_SYNC_BACKUP_OR_RESTORE, "Unexpected zk node {}", zookeeper_path + "/" + zk_node);
-            String host = zk_node.substr(0, separator_pos);
-            String status = zk_node.substr(separator_pos + 1);
-            if (status == kErrorStatus)
-            {
-                host_with_error = host;
-                error_message = zookeeper->get(zookeeper_path + "/" + zk_node);
-                return;
-            }
-            auto it = unready_hosts.find(host);
-            if ((it != unready_hosts.end()) && (status == new_status))
-            {
-                String result = zookeeper->get(zookeeper_path + "/" + zk_node);
-                for (size_t i : it->second)
-                    ready_hosts_results[i] = result;
-                unready_hosts.erase(it);
-            }
-        }
-    };
-
-    /// Wait until all hosts are ready or an error happens or time is out.
-    std::atomic<bool> watch_set = false;
-    std::condition_variable watch_triggered_event;
-
-    auto watch_callback = [&](const Coordination::WatchResponse &)
-    {
-        watch_set = false; /// After it's triggered it's not set until we call getChildrenWatch() again.
-        watch_triggered_event.notify_all();
-    };
-
-    auto watch_triggered = [&] { return !watch_set; };
-
-    bool use_timeout = timeout_ms.has_value();
-    std::chrono::milliseconds timeout{timeout_ms.value_or(0)};
-    std::chrono::steady_clock::time_point start_time = std::chrono::steady_clock::now();
-    std::chrono::steady_clock::duration elapsed;
-    std::mutex dummy_mutex;
-
-    while (!unready_hosts.empty() && !error_message)
-    {
-        watch_set = true;
-        Strings nodes = zookeeper->getChildrenWatch(zookeeper_path, nullptr, watch_callback);
-        process_zk_nodes(nodes);
-
-        if (!unready_hosts.empty() && !error_message)
-        {
-            LOG_TRACE(log, "Waiting for host {}", unready_hosts.begin()->first);
-            std::unique_lock dummy_lock{dummy_mutex};
-            if (use_timeout)
-            {
-                elapsed = std::chrono::steady_clock::now() - start_time;
-                if ((elapsed > timeout) || !watch_triggered_event.wait_for(dummy_lock, timeout - elapsed, watch_triggered))
-                    break;
-            }
-            else
-                watch_triggered_event.wait(dummy_lock, watch_triggered);
-        }
-    }
-
-    if (watch_set)
-    {
-        /// Remove watch by triggering it.
-        zookeeper->create(zookeeper_path + "/remove_watch-", "", zkutil::CreateMode::EphemeralSequential);
-        std::unique_lock dummy_lock{dummy_mutex};
-        watch_triggered_event.wait(dummy_lock, watch_triggered);
-    }
-
-    if (error_message)
-        throw Exception(ErrorCodes::FAILED_TO_SYNC_BACKUP_OR_RESTORE, "Error occurred on host {}: {}", *host_with_error, *error_message);
-
-    if (!unready_hosts.empty())
-    {
-        throw Exception(
-            ErrorCodes::FAILED_TO_SYNC_BACKUP_OR_RESTORE,
-            "Waited for host {} too long ({})",
-            unready_hosts.begin()->first,
-            to_string(elapsed));
-    }
-
-    return ready_hosts_results;
 }
 
 }
